ml-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
requests==2.32.5
python-dotenv==1.0.1
google-generativeai==0.8.3
